fix(auth): review #2 — resolve signup mode server-side, fix dead escape hatch
Review #2 (manual + subagent, full independent re-read) findings, all fixed.

HIGH
- "Live the same day" was gated on the ?intent= URL param, which the most
  common pre-launch arrival doesn't carry: signupCta omits it for the
  `closed` mode, and the header's "Sign in" link is a bare /auth on every
  marketing page. A gated visitor saw "Live the same day", submitted, and
  was told they were in a queue. The param can't answer this — it says
  which button was pressed, not what the platform is doing. Split the
  route into a server shell that resolves cfg_platform_stage.signupMode
  (as components/shared/header.tsx already does) and a client AuthFlow
  that takes it as a prop. ?intent= still drives the heading wording.
- The new "Trouble signing in?" escape hatch pointed at /contact, which
  is NOT in the middleware PUBLIC_PATHS allowlist. /auth is allowlisted
  precisely so approved launch partners can sign in while the coming-soon
  gate is up, so the one group the link exists for would have been
  bounced to the teaser page. Now a mailto.

MEDIUM
- The live region only announced the cooldown expiring. Pressing "Resend
  link" disables the focused button — dropping focus to <body> — so a
  screen-reader user got no confirmation the resend happened at all. The
  region now carries the three real events: sending, sent, and cooldown
  expired.
- Dropping the marketing footer removed Terms/Privacy/Cookies from all
  four auth routes, not just /auth — including /auth/complete-profile,
  which collects personal data. Replaced with a one-line legal strip in
  the layout, which covers all four at a fraction of the height so the
  dark panel still runs nearly full viewport.
- `now` is seeded at mount and only advanced while cooling down, so the
  first render of "sent" compared a fresh cooldownEndsAt against a stale
  clock: a visitor who read for 90s then submitted saw "Resend in 150s"
  until the next tick. Re-anchored on submit. Pre-existing on master, but
  this PR rewrote everything around it.

LOW
- Moved the three-promise row into pillar-console.ts and consumed it from
  the landing hero — it was the exact duplication the module was created
  to prevent, and the promises only work if there is one copy.
- Eyebrow now composes via cn() so a future caller passing a non-colour
  class doesn't silently lose the default colour.
- Straight apostrophe in the pre-launch promise; referral banner missing
  the sm:px-8 its siblings carry; trimmed comments that narrated the diff
  rather than the code.
